pyproject.toml
[project]
name = "artify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
]
